更新查询案例

+++ src/main/java/ewing/user/UserServiceImpl.java
@@ -6,6 +6,7 @@ import com.querydsl.sql.dml.SQLUpdateClause;
 import ewing.application.AppAsserts;
 import ewing.application.common.When;
 import ewing.application.query.Page;
+import ewing.application.query.Where;
 import ewing.dao.entity.Role;
 import ewing.dao.entity.User;
 import ewing.dao.entity.UserRole;
@@ -82,8 +83,7 @@ public class UserServiceImpl implements UserService {
         AppAsserts.notNull(userWithRole.getUserId(), "用户ID不能为空！");
 
         // 更新用户的角色列表
-        userRoleDao.deleteWhere(qUserRole
-                .userId.eq(userWithRole.getUserId()));
+        userRoleDao.deleteWhere(qUserRole.userId.eq(userWithRole.getUserId()));
         addUserRoles(userWithRole);
 
         // 更新用户
@@ -104,13 +104,9 @@ public class UserServiceImpl implements UserService {
     public Page<UserWithRole> findUserWithRole(FindUserParam findUserParam) {
         BooleanExpression expression = Expressions.TRUE;
         // 用户名
-        expression = expression.and(StringUtils.hasText(
-                findUserParam.getUsername()) ? qUser.username
-                .contains(findUserParam.getUsername()) : null);
+        expression = expression.and(Where.hasText(findUserParam.getUsername(),qUser.username::contains));
         // 昵称
-        expression = expression.and(StringUtils.hasText(
-                findUserParam.getNickname()) ? qUser.nickname
-                .contains(findUserParam.getNickname()) : null);
+        expression = expression.and(Where.hasText(findUserParam.getNickname(),qUser.nickname::contains));
         return userDao.findUserWithRole(findUserParam, expression);
     }
